refact addverb
