+++ MapFileDict/MapFileDict.csproj
@@ -57,9 +57,9 @@
     <Compile Include="MapFileList.cs" />
     <Compile Include="MemMap.cs" />
     <Compile Include="MemMapStream.cs" />
-    <Compile Include="MyClassThatRunsIn32and64Bit.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="OutOfProc.cs" />
+    <Compile Include="OutOfProcBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
